[stinkytofu] Add clang-tidy config, CMake modules, and bug fixes (#7526)
## Summary

- Create reusable `cmake/modules/ClangTidy.cmake` and
`CheckToolVersion.cmake` — any project can adopt by setting variables
before `include(ClangTidy)`
- Add `.clang-tidy` config with explicit bug-finding checks only
(bugprone-* + clang-analyzer-*), zero warnings
- Add `invoke tidy` local developer command
- Fix bugs found by clang-tidy

## Details

**New cmake modules** (`cmake/modules/`):
- `ClangTidy.cmake` — provides `add_clang_tidy_custom_target()` and
`clang_tidy_check(TARGET)`. Configurable via `ENABLE_CLANG_TIDY`,
`CLANG_TIDY_SOURCE_FILTER`, `CLANG_TIDY_EXTRA_ARGS`,
`CLANG_TIDY_CONFIG_FILE`
- `CheckToolVersion.cmake` — finds and version-checks clang-tidy-20 /
run-clang-tidy-20

**Bug fixes** (from clang-tidy findings):
- Remove unused variables (`CompositeInstructionLoweringPass.cpp`,
`RedundantMovEliminationPassTest.cpp`, `EstimateAsmCyclesPass.cpp`)
- Fix implicit widening casts (`EstimateAsmCyclesPass.cpp`,
`SwPrefetchInsertionPass.cpp`)
- Add null-pointer assert before dereference (`LegalizationUtils.cpp`)
- Remove dead store (`GenInstructions.cpp`)
- Initialize `RewriteStmt::kind` field to fix uninitialized-field
warning
- Simplify redundant optional access (`RawAsmParser.cpp`)

**Mechanical fixes** (auto-applied with `clang-tidy --fix`):
- `modernize-use-override`, `modernize-use-equals-default`
- `readability-container-contains`, `modernize-use-starts-ends-with`
- `performance-unnecessary-value-param`, `performance-for-range-copy`

## Test plan

- [x] `invoke build --no-python && invoke tidy` — zero warnings locally
- [x] All 569 unit tests pass
- [x] CI clang-tidy workflow will be added in a follow-up PR after this
merges

+++ shared/stinkytofu/src/transforms/asm/SwPrefetchInsertionPass.cpp
@@ -166,7 +166,7 @@ int64_t insertSwPrefetchInstPcRelBefore(
 
     bb.insertIR(anchorIt, movInst);
     bb.insertIR(anchorIt, prefetchInst);
-    return static_cast<int64_t>(movB + pfB);
+    return static_cast<int64_t>(movB) + static_cast<int64_t>(pfB);
 }
 
 /// One mov+prefetch pair before \p anchorIt (`s_getpc_b64`), chaining \p
